requirements.txt
openai>=1.52.0
python-dotenv>=1.0.1
mem0ai>=1.0.4
letta-client
zep-cloud>=3.18.0
scipy>=1.9.0
statsmodels>=0.14.0
